Fix concurrency bug showing up on ARM
There are two places in the code where we use "volatile"
to aid in inter-thread communication.  While this has worked
fine so far on all platforms we tested, bugs popped up in this
code when running it on ARM processors (the Apple M1 in this case).

Generally, x86 CPUs have stricter memory ordering semantics by
default than ARM CPUs (e.g. won't re-order write relative to
other writes etc.). So, it's possible that the prior implementation
was *conditionally* correct given the CPU arch, but it's not correct,
I believe, with respect to what the standard guarantees.

This commit replaces those usages of `volatile` with `std::atomic`.
Since the underlying type this is being applied to is an enum
represented as a simple uint8_t, this was a trivial replacement
and should work well.  In small scale testing, there was no
performance degredation on x86_64 (i.e. on newton).  However, it is
worth noting that the unitig extraction code runs considerably
slower on ARM than on x86 (3-4x slower).  We should look into optimizing
this if we want ARM support to be first-class.