Fix latent bug in vector allocation
The `reserve` function does not change the size of this vector, but simply ensures that *in the future*, there will be sufficient space for at least `buffer_elem_count` elements without the need to resize.

However, we later read directly into this vector and then access it without ever resizing it.  This is UB, though it has mostly compiled and behaved as one might have expected if the vector *had* been resized, probably because `reserve` is doing an eager allocation and we are reading into memory allocated by this process and later reading from that memory.

However, this UB is wrong, and recent versions of Clang are clever enough to realize this and replace access to `prefix_file_buf` directly with the `ud2` opcode, which universally produces an illegal instruction exception.

This change fixes this latent bug, which, seemingly, has existed for years, by ensuring that the vector is properly resized for the elements it will later hold, rather than simply ensuring that the *capacity* is sufficient.